Module Wall.Drawer

Type providing a means to obtain the path from a particular point on the originating face of a wall and the ground.

type loc = [

| `BL

| `BR

| `TL

| `TR

| `CN

| `B of float

| `T of float

| `L of float

| `R of float

| `XY of float * float

]

type t = loc -> OCADml.Path3.t

val translate : OCADml.V3.t -> t -> t

val xtrans : float -> t -> t

val ytrans : float -> t -> t

val ztrans : float -> t -> t

val rotate : ?about:OCADml.V3.t -> OCADml.V3.t -> t -> t

val xrot : ?about:OCADml.V3.t -> float -> t -> t

val yrot : ?about:OCADml.V3.t -> float -> t -> t

val zrot : ?about:OCADml.V3.t -> float -> t -> t

val axis_rotate : ?about:OCADml.V3.t -> OCADml.V3.t -> float -> t -> t

val quaternion : ?about:OCADml.V3.t -> OCADml.Quaternion.t -> t -> t

val scale : OCADml.V3.t -> t -> t

val xscale : float -> t -> t

val yscale : float -> t -> t

val zscale : float -> t -> t

val mirror : OCADml.V3.t -> t -> t

val affine : OCADml.Affine3.t -> t -> t

val map : (OCADml.Path3.t -> OCADml.Path3.t) -> t -> t