Module Walls.Side

type t = Wall.t IMap.t

Maps from index/position to Wall.t, representing a side of the body or thumb of a plate.

val translate : OCADml.V3.t -> t -> t

val xtrans : float -> t -> t

val ytrans : float -> t -> t

val ztrans : float -> t -> t

val rotate : ?about:OCADml.V3.t -> OCADml.V3.t -> t -> t

val xrot : ?about:OCADml.V3.t -> float -> t -> t

val yrot : ?about:OCADml.V3.t -> float -> t -> t

val zrot : ?about:OCADml.V3.t -> float -> t -> t

val axis_rotate : ?about:OCADml.V3.t -> OCADml.V3.t -> float -> t -> t

val quaternion : ?about:OCADml.V3.t -> OCADml.Quaternion.t -> t -> t

val scale : OCADml.V3.t -> t -> t

val xscale : float -> t -> t

val yscale : float -> t -> t

val zscale : float -> t -> t

val mirror : OCADml.V3.t -> t -> t

val affine : OCADml.Affine3.t -> t -> t

type config = int -> Wall.config option

A function taking a index/position along a side of the body or thumb of a plate, and returning some Wall.config if there should be a Wall.t made at that location, or none if not.