Module Dometyl__.Plate

module Lookups : sig ... end

Provides per-column lookup functions to be used by make for placement of KeyHole.ts to form Column.ts, and for arrangment and orientation of the generated columns to form the plate.

type 'k config = {

n_rows : Base.int;

centre_row : Base.int;

n_cols : Base.int;

centre_col : Base.int;

spacing : Base.float;

tent : Base.float;

thumb_offset : Scad_ml.Vec3.t;

thumb_angle : Scad_ml.Vec3.t;

}

Non-function parameters used for the construction of the plate.

type 'k t = {

config : 'k config;

scad : Scad_ml.Model.t;

columns : 'k Dometyl__.Columns.t;

thumb : 'k Dometyl__.Column.t;

}

The plate is constructed from a main body of Columns.t and a thumb cluster, modelled as a separate Column.t. As with lower types, the config as well as the aggregate Model.t are carried around for convenient access.

Basic transformation functions, applied to all relevant non-config contents.

include Dometyl__.Sigs.Transformable' with type 'k t := 'k t

type 'a t

val translate : Scad_ml.Vec3.t -> 'a t -> 'a t

include Dometyl__.Sigs.Rotatable' with type 'a t := 'a t

type 'a t

val rotate : Scad_ml.Vec3.t -> 'a t -> 'a t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> 'a t -> 'a t

val make_thumb : n_keys:Base.int -> centre_idx:Base.int -> curve:'k Dometyl__.Curvature.t -> rotate_clips:Base.bool -> 'k Dometyl.KeyHole.t -> 'k Dometyl__.Column.t

make_thumb ~n_keys ~centre_idx ~curve ~rotate_clips keyhole

Factored out helper function for creation of the thumb cluster, using the distribution specification curve to place n_keys keyholes, around the centre_idx key. rotate_clips flags whether to rotate the switch-clips / sockets / etc by 90 degrees. Non-rotated, the clips will be however they are on a usual column (e.g. oriented top to bottom). You may want to set this according to orientation you would like to have your caps in, since Mx stems are not necessarily radially symmetric.

val make : ?⁠n_rows:Base.int -> ?⁠centre_row:Base.int -> ?⁠n_cols:Base.int -> ?⁠centre_col:Base.int -> ?⁠spacing:Base.float -> ?⁠tent:Base.float -> ?⁠n_thumb_keys:Base.int -> ?⁠thumb_centre:Base.int -> ?⁠thumb_curve:'k Dometyl__.Curvature.t -> ?⁠rotate_thumb_clips:Base.bool -> ?⁠thumb_offset:Scad_ml.Vec3.t -> ?⁠thumb_angle:Scad_ml.Vec3.t -> ?⁠lookups:'k Lookups.t -> 'k Dometyl.KeyHole.t -> 'k t

make ?n_rows ?centre_row ?n_cols ?centre_col ?spacing ?tent ?n_thumb_keys ?thumb_centre ?thumb_curve ?rotate_thumb_clips ?thumb_offset ?thumb_angle ?lookups keyhole

Create a switch plate, with provided optional parameters and defaults. The default parameters are used by the Splaytyl and Closed configurations, and may or may not be suitable as a jumping off point for your own design. To that end, other examples with divergent parameters are provided in the modules within the boards sub-directory.

n_rows specifies number of rows per column

centre_row is used by key distribution functions, namely Curvature.place. Relevant to thumb_curve and Lookups.t.curve.

n_cols specifies number of columns on the main body of the switch plate

spacing gives the distance in mm between each column by default. Further adjustment can of course be done with Lookups.t.offset.

tent angle in radians that the entire plate (including thumb) will be y-rotated

n_thumb_keys, thumb_centre, thumb_curve and rotate_thumb_clips: see make_thumb

thumb_offset directs the translation used for thumb cluster placement (starting point if zero origin).

thumb_angle sets the euler (x -> y -> z) rotation used to orient the thumb before translating to its destination (rotation is about 0., oriented along x to start).

lookups provides various per-column key/column-placement and orientation functions. See Lookups.

keyhole is the KeyHole.t that will be distributed to make up the main plate and thumb cluster, carring all contained coordinate information and accesory Model.ts along with them to their new homes.

val column_joins : 'k t -> Scad_ml.Model.t

column_joins t

Returns a Model.t scad which joins together all columns of the main plate by hulling the KeyHole.Face.t.scads of the KeyHole.ts together, for a closed switch-plate look (and sturdiness). Typically, a function such as this or skeleton_bridges below will be used as as the plate_welder parameter to Case.make.

val skeleton_bridges : 'k t -> Scad_ml.Model.t

skeleton_bridges t

Returns a Model.t scad with "bridges" between the bottom row keys from index to middle, and the top keys for the remaining columns. Similar to the BastardKB skeletyl. This function, as well as column_joins make use of functions found in the Bridges module, if you are wanting to do something different, that would be a good place to get started.

val to_scad : 'k t -> Scad_ml.Model.t

to_scad t

Extracts the Model.t from t.

val collect : f:(key:Base.int -> data:'k Dometyl.KeyHole.t -> Scad_ml.Model.t Base.list -> Scad_ml.Model.t Base.list) -> 'k t -> Scad_ml.Model.t

collect ~f t

Fold over all KeyHole.ts contained within t, extracting Model.ts with f, and unioning them. Used for collect_caps and collect_cutouts.

val collect_caps : 'k t -> Scad_ml.Model.t

collect_caps t

Return a Model.t representing the union between all caps carried by the KeyHole.ts that make up the plate t. Used to design around keycap closeness, and lookout for collisions etc.

val collect_cutouts : 'k t -> Scad_ml.Model.t

collect_cutout t

Return a Model.t representing the union between all KeyHole.t.cutouts found in the plate t. This is used by Case.make to perform the clearance cut from the Case.t.scad model to make room for whatever the KeyHole.t.cutout is modelling (e.g. hotswap sockets). Like collect_caps, this could also be useful for visualization.