Module Generator.Column

module Join : sig ... end

A hull between KeyHole.ts of a column, and thin scads for each of the outer faces (available for hulling between columns in Bridge.).

type 'k config = {

key : 'k KeyHole.t;

n_keys : Base.int;

curve : Base.int -> 'k KeyHole.t -> 'k KeyHole.t;

}

type 'k t = {

config : 'k config;

scad : Scad_ml.Scad.d3;

keys : 'k KeyHole.t Base.Map.M(Base.Int).t;

joins : Join.t Base.Map.M(Base.Int).t;

}

val translate : Scad_ml.Vec3.t -> 'k t -> 'k t

val scale : Scad_ml.Vec3.t -> 'k t -> 'k t

val rotate : Scad_ml.Vec3.t -> 'k t -> 'k t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> 'k t -> 'k t

val quaternion : Scad_ml.Quaternion.t -> 'k t -> 'k t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> 'k t -> 'k t

val mirror : Scad_ml.Vec3.t -> 'k t -> 'k t

val make : ?join_ax:[< `EW | `NS NS ] -> n_keys:Base.int -> curve:(Base.int -> 'k KeyHole.t -> 'k KeyHole.t) -> caps:(Base.int -> Scad_ml.Scad.d3) -> 'k KeyHole.t -> 'k t

make ?join_ax ~n_keys ~curve ~caps key

Create a column by distributing n_keys copies of key using the distribution function curve, which will position each KeyHole.t according to the index (from 0 to n_keys - 1), before they are joined via Scad.hull between the KeyHole.t.faces along the axis indicated by join_ax. The caps lookup places a key cap above each KeyHole.t corresponding to the index (examples/defaults can be found in Caps).

val to_scad : 'k t -> Scad_ml.Scad.d3