Module Generator.Connect

type t = {

scad : Scad_ml.Scad.d3;

outline : Scad_ml.Vec3.t Base.list;

inline : Scad_ml.Vec3.t Base.list;

}

Shape(s) connecting walls together, and the points along the outside and inside of the foot.

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

type config

val clockwise_union : t Base.list -> t

clockwise_union ts Create a union of ts. They should be provided in clockwise order, such that the outline and inline of the resulting t are continous and in the clockwise direction.

val outline_2d : t -> (Base.float * Base.float) Base.list

outline_2d t

Retrieve the outline points of t and project them to 2d vectors.

val inline_2d : t -> (Base.float * Base.float) Base.list

inline_2d t

Retrieve the inline points of t and project them to 2d vectors.

val straight : ?n_facets:Base.int -> ?height:Base.float -> ?fudge_factor:Base.float -> ?overlap_factor:Base.float -> ?min_width:Base.float -> Base.unit -> config

val bez : ?n_facets:Base.int -> ?height:Base.float -> ?n_steps:Base.int -> Base.unit -> config

val cubic : ?n_facets:Base.int -> ?height:Base.float -> ?scale:Base.float -> ?d:Base.float -> ?n_steps:Base.int -> ?bow_out:Base.bool -> Base.unit -> config

val snake : ?n_facets:Base.int -> ?height:Base.float -> ?scale:Base.float -> ?d:Base.float -> ?n_steps:Base.int -> Base.unit -> config

val full_join : ?n_steps:Wall.Steps.t -> ?fudge_factor:Base.float -> ?overlap_factor:Base.float -> Base.unit -> config

val elbow : ?n_facets:Base.int -> ?height:Base.float -> ?n_steps:Base.int -> ?d:Base.float -> Base.unit -> config

val manual : ?west:(Base.int -> config) -> ?north:(Base.int -> config) -> ?south:(Base.int -> config) -> ?east:(Base.int -> config) -> ?east_link:config -> ?thumb_east:(Base.int -> config) -> ?thumb_south:(Base.int -> config) -> ?thumb_west:(Base.int -> config) -> ?thumb_north:(Base.int -> config) -> ?west_link:config -> Walls.t -> t

val skeleton : ?n_facets:Base.int -> ?index_height:Base.float -> ?height:Base.float -> ?min_straight_width:Base.float -> ?n_steps:Base.int -> ?body_join_steps:Wall.Steps.t -> ?thumb_join_steps:Wall.Steps.t -> ?fudge_factor:Base.float -> ?join_fudge_factor:Base.float -> ?overlap_factor:Base.float -> ?cubic_d:Base.float -> ?cubic_scale:Base.float -> ?thumb_height:Base.float -> ?east_link:config -> ?west_link:config -> ?north_joins:(Base.int -> Base.bool) -> ?south_joins:(Base.int -> Base.bool) -> ?pinky_idx:Base.int -> ?pinky_elbow:Base.bool -> ?close_thumb:Base.bool -> Walls.t -> t

val closed : ?body_steps:Wall.Steps.t -> ?thumb_steps:Wall.Steps.t -> ?fudge_factor:Base.float -> ?overlap_factor:Base.float -> ?west_link:config -> ?east_link:config -> Walls.t -> t

val to_scad : t -> Scad_ml.Scad.d3