Module Generator.Plate

Switch plate generation.

module Lookups : sig ... end

Provides per-column lookup functions to be used by make for placement of KeyHole.ts to form Column.ts, and for arrangment and orientation of the generated columns to form the plate.

type 'k config = {

n_body_rows : Base.int -> Base.int;

body_centres : Base.int -> Base.float;

n_body_cols : Base.int;

centre_col : Base.int;

spacing : Base.float;

tent : Base.float;

n_thumb_rows : Base.int -> Base.int;

thumb_centres : Base.int -> Base.float;

n_thumb_cols : Base.int;

thumb_offset : Scad_ml.Vec3.t;

thumb_angle : Scad_ml.Vec3.t;

}

Non-function parameters used for the construction of the plate.

type 'k t = {

config : 'k config;

scad : Scad_ml.Scad.d3;

body : 'k Columns.t;

thumb : 'k Columns.t;

}

The plate is constructed from a main body of Columns.t and a thumb cluster, modelled as a separate Column.t. As with lower types, the config as well as the aggregate Scad.d3 are carried around for convenient access.

val translate : Scad_ml.Vec3.t -> 'k t -> 'k t

val scale : Scad_ml.Vec3.t -> 'k t -> 'k t

val rotate : Scad_ml.Vec3.t -> 'k t -> 'k t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> 'k t -> 'k t

val quaternion : Scad_ml.Quaternion.t -> 'k t -> 'k t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> 'k t -> 'k t

val mirror : Scad_ml.Vec3.t -> 'k t -> 'k t

val make : ?n_body_cols:Base.int -> ?centre_col:Base.int -> ?spacing:Base.float -> ?tent:Base.float -> ?n_thumb_cols:Base.int -> ?rotate_thumb_clips:Base.bool -> ?thumb_offset:Scad_ml.Vec3.t -> ?thumb_angle:Scad_ml.Vec3.t -> ?body_lookups:'k Lookups.t -> ?thumb_lookups:'k Lookups.t -> ?caps:(Base.int -> Scad_ml.Scad.d3) -> ?thumb_caps:(Base.int -> Scad_ml.Scad.d3) -> 'k KeyHole.t -> 'k t

make ?n_rows ?centre_row ?n_cols ?centre_col ?spacing ?tent ?n_thumb_keys ?thumb_centre ?thumb_curve ?rotate_thumb_clips ?thumb_offset ?thumb_angle ?lookups ?caps ?thum_caps keyhole

Create a switch plate, with provided optional parameters and defaults. The default parameters are used by the Splaytyl and Closed configurations, and may or may not be suitable as a jumping off point for your own design. To that end, other examples with divergent parameters are provided in the modules within the boards sub-directory.

n_cols specifies number of columns on the main body of the switch plate

centre_col specifies the column around which the others are positioned (typically 2, the middle finger) on the main body of the switch plate

spacing gives the distance in mm between each column by default. Further adjustment can of course be done with Lookups.t.offset.

tent angle in radians that the entire plate (including thumb) will be y-rotated

n_thumb_keys, thumb_centre, thumb_curve and rotate_thumb_clips: see make_thumb

thumb_offset directs the translation used for thumb cluster placement (starting point if zero origin).

thumb_angle sets the euler (x -> y -> z) rotation used to orient the thumb before translating to its destination (rotation is about 0., oriented along x to start).

lookups provides various per-column key/column-placement and orientation functions. See Lookups.

caps is a lookup from row number to a Scad.d3 representing a keycap. This will be used to provide caps to KeyHole.ts before they are distributed in Column.make.

thumb_caps same as caps, but for the thumb. If not provided, this default to caps.

keyhole is the KeyHole.t that will be distributed to make up the main plate and thumb cluster, carring all contained coordinate information and accesory Scad.d3s along with them to their new homes.

val column_joins : ?in_d:Base.float -> ?out_d1:Base.float -> ?out_d2:Base.float -> 'k t -> Scad_ml.Scad.d3

column_joins ?d1 ?d2 t

Returns a Scad.d3 scad which joins together all columns of the main plate by hulling the KeyHole.Face.t.scads of the KeyHole.ts together, for a closed switch-plate look (and sturdiness). Typically, a function such as this or skeleton_bridges below will be used as as the plate_welder parameter to Case.make. Optional params in_d, out_d1 and out_d2 are passed on to Bridges.cols to control how far the lower key's face is moved out, and the upper keys face is moved in before being hulled with its neighbours across the gap. This can be used to thicken the generated supports.

val skeleton_bridges : ?in_d:Base.float -> ?out_d1:Base.float -> ?out_d2:Base.float -> 'k t -> Scad_ml.Scad.d3

skeleton_bridges ?d1 ?d2 t

Returns a Scad.d3 scad with "bridges" between the bottom row keys from index to middle, and the top keys for the remaining columns. Similar to the BastardKB skeletyl. This function, as well as column_joins make use of functions found in the Bridges module, if you are wanting to do something different, that would be a good place to get started. Optional params in_d, out_d1 and out_d2 are passed on to Bridges.cols to control how far the lower key's face is moved out, and the upper keys face is moved in before being hulled with its neighbours across the gap. This can be used to thicken the generated supports.

val to_scad : 'k t -> Scad_ml.Scad.d3

to_scad t

Extracts the Scad.d3 from t.

val collect : f:(key:Base.int -> data:'k KeyHole.t -> Scad_ml.Scad.d3 Base.list -> Scad_ml.Scad.d3 Base.list) -> 'k t -> Scad_ml.Scad.d3

collect ~f t

Fold over all KeyHole.ts contained within t, extracting Scad.d3s with f, and unioning them. Used for collect_caps and collect_cutouts.

val collect_caps : 'k t -> Scad_ml.Scad.d3

collect_caps t

Return a Scad.d3 representing the union between all caps carried by the KeyHole.ts that make up the plate t. Used to design around keycap closeness, and lookout for collisions etc.

val collect_cutouts : 'k t -> Scad_ml.Scad.d3

collect_cutout t

Return a Scad.d3 representing the union between all KeyHole.t.cutouts found in the plate t. This is used by Case.make to perform the clearance cut from the Case.t.scad model to make room for whatever the KeyHole.t.cutout is modelling (e.g. hotswap sockets). Like collect_caps, this could also be useful for visualization.