Module Wall.Edges

Bezier curves representing the four edges running from the start to the foot of each Wall.t

type t = {

top_left : Edge.t;

top_right : Edge.t;

bot_left : Edge.t;

bot_right : Edge.t;

}

Record containting an Edge.t running from the corners of KeyHole.Face.t.points down to the floor, where the terminal points are used for Wall.t.foot.

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val map : f:(Edge.t -> Edge.t) -> t -> t

val of_clockwise_list_exn : Edge.t Base.list -> t

of_clockwise_list_exn l Convert a four element list into a t. The ordering actually shouldn't just be clockwise, but is assumed to literally be: TL, TR, BR, BL.

val of_clockwise_list : Edge.t Base.list -> (t, Base.string) Base.Result.t

val get : t -> [< `BL | `BR | `TL | `TR ] -> Edge.t

get t corner

Access fields from record t according to provided corner tag.