Module Wall.Steps

Configuration type for number of steps used to draw the bezier curves that form the walls.

type t = [

| `Flat of Base.int

| `PerZ of Base.float

]

`Flat n simply indicates n steps, `PerZ mm specifies that there should be a step for every mm off the ground the start of the wall is.

val to_int : [< `Flat of Base.int | `PerZ of Base.float ] -> Base.float -> Base.int

to_int t z

Converts to a discrete number of steps. In the case of `Flat, it is simply the contained integer and the z value will be ignored. For `PerZ, the resulting number of steps will be how many times the wrapped float fits into the provided z value.