Module Generator.Walls

type presence =

| No

| Yes

| Eye

Used for lookups, indicating whether to place a wall at a particular position on the plate, and if so, whether there should be an eyelet attached.

module Side : sig ... end

module Sides : sig ... end

val auto_body : ?d1:Base.float -> ?d2:Base.float -> ?z_off:Base.float -> ?thickness:Base.float -> ?index_thickness:Base.float -> ?north_clearance:Base.float -> ?south_clearance:Base.float -> ?side_clearance:Base.float -> ?n_steps:Wall.Steps.t -> ?n_facets:Base.int -> ?north_lookup:(Base.int -> presence) -> ?south_lookup:(Base.int -> presence) -> ?west_lookup:(Base.int -> presence) -> ?east_lookup:(Base.int -> presence) -> ?eyelet_config:Eyelet.config -> 'k Plate.t -> Sides.t

val auto_thumb : ?d1:Base.float -> ?d2:Base.float -> ?z_off:Base.float -> ?thickness:Base.float -> ?index_thickness:Base.float -> ?north_clearance:Base.float -> ?south_clearance:Base.float -> ?side_clearance:Base.float -> ?n_steps:Wall.Steps.t -> ?n_facets:Base.int -> ?north_lookup:(Base.int -> presence) -> ?south_lookup:(Base.int -> presence) -> ?west_lookup:(Base.int -> presence) -> ?east_lookup:(Base.int -> presence) -> ?eyelet_config:Eyelet.config -> 'k Plate.t -> Sides.t

type t = {

body : Sides.t;

thumb : Sides.t;

}

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val make : body:Sides.t -> thumb:Sides.t -> t

val manual : body_west:Side.config -> body_north:Side.config -> body_east:Side.config -> body_south:Side.config -> thumb_south:Side.config -> thumb_north:Side.config -> thumb_east:Side.config -> thumb_west:Side.config -> 'k Plate.t -> t

val to_scad : t -> Scad_ml.Scad.d3

val collect_screws : t -> Eyelet.t Base.list