Module Walls.Sides

type t = {

west : Side.t;

north : Side.t;

east : Side.t;

south : Side.t;

}

Four cardinal Side.ts for a body/thumb.

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val manual_body : ?spacing:Base.float -> west:Side.config -> north:Side.config -> east:Side.config -> south:Side.config -> 'k Columns.t -> t

manual_body ?spacing ~west ~north ~east ~south columns

Construct a t with the given Side.configs for each of the edges of the provided Plate.t body columns (columns oriented along the Y-axis (N-S)). ?spacing is passed along to Wall.column_drop, and should match the spacing config parameter of the Plate.t that columns belongs to.

val manual_thumb : west:Side.config -> north:Side.config -> east:Side.config -> south:Side.config -> 'k Columns.t -> t

manual_thumb ~west ~north ~east ~south columns

Construct a t with the given Side.configs for each of the edges of the provided Plate.t thumb columns (columns oriented along the X-axis (W-E)). Note that the relevant indices are increment in directions reflecting the 90 degree rotation of the thumb columns relative to the body.

val auto : ?d1:Base.float -> ?d2:Base.float -> ?z_off:Base.float -> ?thickness:Base.float -> ?index_thickness:Base.float -> ?north_clearance:Base.float -> ?south_clearance:Base.float -> ?side_clearance:Base.float -> ?n_steps:Wall.Steps.t -> ?n_facets:Base.int -> ?north_lookup:( Base.int -> presence ) -> ?south_lookup:( Base.int -> presence ) -> ?west_lookup:( Base.int -> presence ) -> ?east_lookup:( Base.int -> presence ) -> ?eyelet_config:Eyelet.config -> ?spacing:Base.float -> ?thumb:Base.bool -> 'k Columns.t -> t

auto ?d1 ?d2 ?z_off ?thickness ?index_thickness ?north_clearance ?south_clearance ?side_clearance ?n_steps ?n_faces ?north_lookup ?south_lookup ?west_lookup ?east_lookup ?eyelet_config ?spacing ?thumb columns

Construct a t from a collection of parameters to be applied consistently across the generated Wall.ts.

d1, d2, z_off, thickness, n_steps, n_facets, and eyelet_config are as described in the documentation of Wall.poly_siding

index_thickness can be used to override thickness for the first two columns of the body.

{north,south,side}_clearance provide clearance to Wall.poly_siding for the corresponding sections.

spacing is given to Wall.column_drop for north and south walls when thumb is false

thumb flags whether columns represents a thumb plate (default is false)

{north,south,west,east}_lookup parameters are functions from index to presence, indicating whether there is a wall present at a position or not (and if there should be an eyelet attached to its base)

val get : t -> [ `N | `E | `S | `W ] -> Side.t

val to_scad : t -> Scad_ml.Scad.d3

val collect_screws : ?init:Eyelet.t Base.list -> t -> Eyelet.t Base.list