Module Generator__.Walls

type presence =

| No

| Yes

| Eye

Used for lookups, indicating whether to place a wall at a particular position on the plate, and if so, whether there should be an eyelet attached.

module Body : sig ... end

Generation and storage of Wall.ts from the KeyHole.ts of the main body of the plate (excluding the thumb cluster).

module Thumb : sig ... end

type t = {

body : Body.t;

thumb : Thumb.t;

}

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val to_scad : t -> Scad_ml.Scad.t

val collect_screws : t -> Generator.Eyelet.t Base.list