Module Generator__Wall

module Steps : sig ... end

Configuration type for number of steps used to draw the bezier curves that form the walls.

module Edge : sig ... end

Bezier curve wall edge type and helpers.

module Edges : sig ... end

Bezier curves representing the four edges running from the start to the foot of each Wall.t

type t = {

scad : Scad_ml.Model.t;
Aggregate scad, including screw outshoot if included

start : Generator.Points.t;
Corner points of the KeyHole.Face.t this wall emerged from

foot : Generator.Points.t;
Terminal points where the wall meets the XY plane.

edges : Edges.t;
Bezier curves that specify the edge vertices.

screw : Generator.Screw.t Base.option;
Model, coordinates, and config of screw offshoot if included.

}

Record representing a wall extending from a KeyHole.Face.t to the ground.

Basic transformation functions

include Generator.Sigs.Transformable with type t := t

type t

val translate : Scad_ml.Vec3.t -> t -> t

include Generator.Sigs.Rotatable with type t := t

type t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val swing_face : ?⁠step:Base.float -> Scad_ml.Vec3.t -> Generator.KeyHole.Face.t -> Generator.KeyHole.Face.t * Scad_ml.Vec3.t

swing_face ?step key_orgin face

Iteratively find a rotation around faces bottom or top axis, depending on which way it is pointing in z (determined with key_origin), that brings face to a more vertical orientation, returning a pivoted KeyHole.Face.t and it's new orthogonal Vec3.t.

val poly_siding : ?⁠x_off:Base.float -> ?⁠y_off:Base.float -> ?⁠z_off:Base.float -> ?⁠clearance:Base.float -> ?⁠n_steps:[< `Flat of Base.int | `PerZ of Base.float Flat ] -> ?⁠n_facets:Base.int -> ?⁠d1:Base.float -> ?⁠d2:Base.float -> ?⁠thickness:Base.float -> ?⁠screw_config:Generator.Screw.config -> [< `East | `North | `South | `West ] -> 'a Generator.KeyHole.t -> t

val column_drop : ?⁠z_off:Base.float -> ?⁠clearance:Base.float -> ?⁠d1:Base.float -> ?⁠d2:Base.float -> ?⁠thickness:Base.float -> ?⁠n_steps:[< `Flat of Base.int | `PerZ of Base.float Flat ] -> ?⁠n_facets:Base.int -> ?⁠screw_config:Generator.Screw.config -> spacing:Base.float -> columns:'k Generator.Columns.t -> [< `North | `South ] -> Base.int -> t

val start_direction : t -> Scad_ml.Vec3.t

start_direction t

Direction vector from right to left of the wall start points.

val foot_direction : t -> Scad_ml.Vec3.t

foot_direction t

Direction vector from right to left of the wall foot points.

val to_scad : t -> Scad_ml.Model.t