Module Generator__Wall.EdgeDrawer

Functions that find the point along the top and bottom edges of the start position of the wall closest to the given xy position, and provides a bezier edge to the ground starting from there.

type drawer = Scad_ml.Vec3.t -> Edge.t

type t = {

top : drawer;

bot : drawer;

}

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val make : ?⁠max_iter:Base.int -> ?⁠tolerance:Base.float -> get_bez:(Base.bool -> Scad_ml.Vec3.t -> Edge.t) -> Generator.Points.t -> t

val map : f:(drawer -> drawer) -> t -> t