Module Generator__Walls

type presence =

| No

| Yes

| Eye

Used for lookups, indicating whether to place a wall at a particular position on the plate, and if so, whether there should be an eyelet attached.

module Side : sig ... end

module Sides : sig ... end

val make_body : ?⁠d1:Base.float -> ?⁠d2:Base.float -> ?⁠z_off:Base.float -> ?⁠thickness:Base.float -> ?⁠index_thickness:Base.float -> ?⁠north_clearance:Base.float -> ?⁠south_clearance:Base.float -> ?⁠side_clearance:Base.float -> ?⁠n_steps:Generator.Wall.Steps.t -> ?⁠n_facets:Base.int -> ?⁠north_lookup:(Base.int -> presence) -> ?⁠south_lookup:(Base.int -> presence) -> ?⁠west_lookup:(Base.int -> presence) -> ?⁠east_lookup:(Base.int -> presence) -> ?⁠eyelet_config:Generator.Eyelet.config -> 'k Generator.Plate.t -> Sides.t

val make_thumb : ?⁠d1:Base.float -> ?⁠d2:Base.float -> ?⁠z_off:Base.float -> ?⁠thickness:Base.float -> ?⁠index_thickness:Base.float -> ?⁠north_clearance:Base.float -> ?⁠south_clearance:Base.float -> ?⁠side_clearance:Base.float -> ?⁠n_steps:Generator.Wall.Steps.t -> ?⁠n_facets:Base.int -> ?⁠north_lookup:(Base.int -> presence) -> ?⁠south_lookup:(Base.int -> presence) -> ?⁠west_lookup:(Base.int -> presence) -> ?⁠east_lookup:(Base.int -> presence) -> ?⁠eyelet_config:Generator.Eyelet.config -> 'k Generator.Plate.t -> Sides.t

type t = {

body : Sides.t;

thumb : Sides.t;

}

val translate : Scad_ml.Vec3.t -> t -> t

val scale : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val quaternion : Scad_ml.Quaternion.t -> t -> t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val manual : body_west:(Base.int -> Generator.Wall.config Base.option) -> body_north:(Base.int -> Generator.Wall.config Base.option) -> body_east:(Base.int -> Generator.Wall.config Base.option) -> body_south:(Base.int -> Generator.Wall.config Base.option) -> thumb_south:(Base.int -> Generator.Wall.config Base.option) -> thumb_north:(Base.int -> Generator.Wall.config Base.option) -> thumb_east:(Base.int -> Generator.Wall.config Base.option) -> thumb_west:(Base.int -> Generator.Wall.config Base.option) -> 'k Generator.Plate.t -> t

val to_scad : t -> Scad_ml.Scad.d3

val collect_screws : t -> Generator.Eyelet.t Base.list