Module Plate.Lookups

Provides per-column lookup functions to be used by make for placement of KeyHole.ts to form Column.ts, and for arrangment and orientation of the generated columns to form the plate.

type 'k t = {

offset : Base.int -> Scad_ml.Vec3.t;	(*
Translation of columns following default spacing and tenting.
*)

curve : Base.int -> 'k Curvature.t;	(*
Distribution of KeyHole.ts to form the columns
*)

swing : Base.int -> Base.float;	(*
Rotation around columnar y-axis, which can emulate row curvature, adding to or subtracting from rotation due to tenting.
*)

splay : Base.int -> Base.float;	(*
Rotation around columnar z-axis (origin of centre_row key). Applying splay will always have to go hand in hand with adjustments to offset to end up with something sensible.
*)

rows : Base.int -> Base.int;	(*
Number of keys to distribute for the given column.
*)

centre : Base.int -> Base.float;	(*
Relative centre key index for the given column. Used by key distribution functions, namely Curvature.place. Relevant to thumb_curve and Lookups.t.curve.
*)

}

Record containing the column generation and placement functions. When each column is being generated (and placed) its index will be given to the lookup functions here to get the relevant settings.

val default_offset : Base.int -> Scad_ml.Vec3.t

val default_curve : Base.int -> 'k Curvature.t

val default_swing : Base.int -> Base.float

val default_splay : Base.int -> Base.float

val default_rows : Base.int -> Base.int

val default_centre : Base.int -> Base.float

val body : ?offset:(Base.int -> Scad_ml.Vec3.t) -> ?curve:(Base.int -> 'k Curvature.t) -> ?swing:(Base.int -> Base.float) -> ?splay:(Base.int -> Base.float) -> ?rows:(Base.int -> Base.int) -> ?centre:(Base.int -> Base.float) -> Base.unit -> 'k t

body ?offset ?curve ?swing ?splay ()

Optionally provide each of the functions to overide the defaults. See t for overview of what each one is used for.

val thumb : ?offset:(Base.int -> Scad_ml.Vec3.t) -> ?curve:(Base.int -> 'k Curvature.t) -> ?swing:(Base.int -> Base.float) -> ?splay:(Base.int -> Base.float) -> ?rows:(Base.int -> Base.int) -> ?centre:(Base.int -> Base.float) -> Base.unit -> 'k t