Module Generator__Wall

module Steps : sig ... end

Configuration type for number of steps used to draw the bezier curves that form the walls.

module Edge : sig ... end

Bezier curve wall edge type and helpers.

module Edges : sig ... end

Bezier curves representing the four edges running from the start to the foot of each Wall.t

type t = {

scad : Scad_ml.Model.t;
Aggregate scad, including screw outshoot if included

start : Generator.Points.t;
Corner points of the KeyHole.Face.t this wall emerged from

foot : Generator.Points.t;
Terminal points where the wall meets the XY plane.

edges : Edges.t;
Bezier curves that specify the edge vertices.

screw : Generator.Screw.t Base.option;
Model, coordinates, and config of screw offshoot if included.

}

Record representing a wall extending from a KeyHole.Face.t to the ground.

Basic transformation functions

include Generator.Sigs.Transformable with type t := t

type t

val translate : Scad_ml.Vec3.t -> t -> t

val mirror : Scad_ml.Vec3.t -> t -> t

val rotate : Scad_ml.Vec3.t -> t -> t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

val swing_face : ?⁠step:Base.float -> Scad_ml.Vec3.t -> Generator.KeyHole.Face.t -> Generator.KeyHole.Face.t * Scad_ml.Vec3.t

swing_face ?step key_orgin face

Iteratively find a rotation around faces bottom or top axis, depending on which way it is pointing in z (determined with key_origin), that brings face to a more vertical orientation, returning a pivoted KeyHole.Face.t and it's new orthogonal Vec3.t.

val poly_siding : ?⁠x_off:Base.float -> ?⁠y_off:Base.float -> ?⁠z_off:Base.float -> ?⁠clearance:Base.float -> ?⁠n_steps:[< `Flat of Base.int | `PerZ of Base.float Flat ] -> ?⁠n_facets:Base.int -> ?⁠d1:Base.float -> ?⁠d2:Base.float -> ?⁠thickness:Base.float -> ?⁠screw_config:Generator.Screw.config -> [< `East | `North | `South | `West ] -> 'a Generator.KeyHole.t -> t

poly_siding ?x_off ?y_off ?z_off ?clearance ?n_steps ?n_facets ?d1 ?d2 ?thickness ?screw_config side keyhole

Generate a t using an OpenScad polyhedron, drawn from a set of bezier curves from the side facing edges of keyhole. Optional parameters influence the shape of the generated wall:

x_off and y_off shift the target endpoints (on the ground) of the wall

z_off shifts the second bezier control point in z (positive would lead to more arc)

clearance moves the start of the wall out from the face of the keyhole

n_steps controls the number of points used to draw the wall (see: Steps). This impact the aeshetics of the wall, but it also determines how well the wall follows the bezier curve, which can have implications for positioning the cutouts for ports near the bottom of the case. A number of steps that is too high can sometimes cause the generated polyhedrons to fail, as points can bunch up, leading to a mesh that is difficult for the OpenScad engine (CGAL) to close. When this happens, either decreasing number of steps (can be done preferentially for short walls with `PerZ), or increasing n_facets to increase the number of (and decrease the size of) triangles that CGAL can use to close the wall shape.

n_facets sets the number of polyhedron faces assigned to the outside and inside of the wall. Values above one will introduce additional beziers of vertices spaced between the ends of the wall, leading to a finer triangular mesh.

d1 and d2 set the distance projected outward along the orthogonal of the side of the keyhole on the xy-plane used to for the second and third quadratic bezier control points respectively.

thickness influences the thickness of the wall (from inside to outside face)

If provided, screw_config describes the screw/bumpon eyelet that should be added to the bottom of the generated wall.

val column_drop : ?⁠z_off:Base.float -> ?⁠clearance:Base.float -> ?⁠n_steps:[< `Flat of Base.int | `PerZ of Base.float Flat ] -> ?⁠n_facets:Base.int -> ?⁠d1:Base.float -> ?⁠d2:Base.float -> ?⁠thickness:Base.float -> ?⁠screw_config:Generator.Screw.config -> spacing:Base.float -> columns:'k Generator.Columns.t -> [< `North | `South ] -> Base.int -> t

column_drop ?z_off ?clearance ?n_steps ?n_facets ?d1 ?d2 ?thickness ?screw_config ~spacing ~columns idx

Wrapper function for poly_siding specifically for (north and south) column end walls. Unlike poly_siding, which takes a KeyHole.t, this takes the map columns, and an idx specifying the column to generate the wall for. Overhang over the next column (to the right) that may have been introduced by tenting is checked for, and an x offset that will reclaim the desired column spacing is calculated and passed along to poly_siding.

val start_direction : t -> Scad_ml.Vec3.t

start_direction t

Direction vector from right to left of the wall start points.

val foot_direction : t -> Scad_ml.Vec3.t

foot_direction t

Direction vector from right to left of the wall foot points.

val to_scad : t -> Scad_ml.Model.t